backend/requirements.txt
# Locked for Python 3.11.9 on Windows
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.6.3
motor==3.3.2
beanie==1.25.0
python-jose==3.3.0
cryptography==42.0.8
passlib==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1
pydantic==2.7.4
email-validator==2.1.1
python-multipart==0.0.9
aiofiles==23.2.1
groq==0.9.0
tavily-python==0.3.3
httpx==0.27.0
